// Shared/Merchandizer/src/app/models/buyers.model.ts
export class Buyers{
    buyerId: number;
    buyerCode: number;
    name: string;
    mobileNo: string;
    email: string;
    house: string;
    road: string;
    sector: string;
    vatRegNo: string;
    binNo: number;
    eTin: string;
    nid: number;
    tds: boolean;
    upazila: string;
    district: string;
}    